update order method, send email confirm bill
- update entity bill: replace idUser = idAddress
- update entity users: remove address column
- update order method: add page shipping + payment
- send email confirm order
- view, cancel order
- fix error, update project

+++ tmanga/src/main/java/com/website/springmvc/libs/GioHang.java
@@ -83,7 +83,7 @@ public class GioHang {
 	public Item getItemForId(Long id) {
 		Item item = new Item();
 		for (Item i : cart.values()) {
-			if(i.comic.getId() == id) {
+			if((long) i.comic.getId() == (long)id) {
 				item = i;
 			}
 		}
